perf(web): various backend changes
moved some functions around

change some endpoint urls for consistency

improved various html loading routes
