+++ src/bedrock_server_manager/web/routes/main_routes.py
@@ -69,3 +69,12 @@ def task_scheduler_route(server_name: str) -> Response:
             "warning",
         )
         return redirect(url_for(".index"))
+
+
+@main_bp.route("/server/<string:server_name>/monitor")
+@login_required
+def monitor_server_route(server_name: str) -> Response:
+    """Renders the server monitoring page for a specific server."""
+    identity = get_current_identity()
+    logger.info(f"User '{identity}' accessed monitor page for server '{server_name}'.")
+    return render_template("monitor.html", server_name=server_name)
