+++ src/bedrock_server_manager/web/templates/monitor.html
@@ -1,4 +1,4 @@
-<!-- bedrock_server_manager//web/templates/monitor.html -->
+<!-- bedrock_server_manager/web/templates/monitor.html -->
 {# Extends the base layout template #}
 {% extends "base.html" %}
